requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
coverage==6.3.2
decorator==5.1.1
Django==3.0
djangorestframework==3.13.1
executing==0.8.3
install==1.3.5
ipdb==0.13.9
ipython==8.2.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.1.1
wcwidth==0.2.5
